requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
scikit-learn
pandas
requests
python-dotenv